requirement.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.0.7
django-adminrestrict==2.0.2
django-filter==2.3.0
django-taggit==1.3.0
django-tinymce4-lite==1.8.0
Jinja2==2.10.3
jsmin==2.2.2
MarkupSafe==1.1.1
Pillow==6.2.1
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.1.0
boto3=1.14.37
django-storages==1.9.1
boto3==1.9.96
botocore==1.12.96
certifi==2018.10.15
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1
django-crispy-forms==1.7.2
django-heroku==0.3.1

docutils==0.14
gunicorn==19.9.0
idna==2.7
jmespath==0.9.3
Pillow==5.2.0
psycopg2==2.7.7
python-dateutil==2.8.0
pytz==2018.5
requests==2.19.1
s3transfer==0.2.0
six==1.12.0
urllib3==1.23
whitenoise==4.1.2
